Redirect to right lang on simple pages + exhibits
Depends on naming slugs for simple pages and exhibits according to
the convention that Chinese slugs are the same as their English
equivalent except "-zh_cn" is appended. If the language is changed
by the user while on and exhibit or simple page, they will be
redirected to the other language version of that page.

Without these changes the language would be changed but the previous
language's content would still be displayed.