Remove all locale changing code
This happens in Multilanguage and essentially has to because
translations are done based on what the Multilanguage plugin
records as the locale_code and we cannot change that from here.
At least this maintains a clean diff in our Multilanguage fork
that essentially only add ones function and calls it.